+++ base/src/sgpp/base/datatypes/DataVector.cpp
@@ -99,7 +99,7 @@ DataVector DataVector::fromString(const std::string& serializedVector) {
       //      double value = std::stod(&(serializedVector[i]), &next);
       double value = std::atof(&(serializedVector[i]));
 #else
-      float value = std::stof(&(serializedVector[i]));
+      float value = std::atof(&(serializedVector[i]));
 #endif
       v.append(value);
       state = PARSER_STATE::COMMAEND;
